+++ app/(home)/socials.tsx
